requirements.txt
# Core data & network
networkx>=3.3
geopandas>=0.14
pandas>=2.2
numpy>=1.26
scipy>=1.13

# Visualisation
folium>=0.16
matplotlib>=3.9

# AI / Agents
langgraph>=0.2
langchain-anthropic>=0.1

# Data collection
requests>=2.32

# Infrastructure
loguru>=0.7
python-dotenv>=1.0
sqlalchemy>=2.0

# Reports
reportlab>=4.2
